Address code review feedback - simplify int/round calls
- Remove redundant int() wrapping around math.ceil() in PirateDailyText.py
- Simplify int(round(x, 0)) to just round(x) in PirateWeeklyText.py

Both functions already return integers, so the extra wrapping was unnecessary.

+++ API/PirateDailyText.py
@@ -1327,7 +1327,7 @@ def calculate_day_text(
         if total_snow_error <= 0:
             snow_sentence = [
                 "centimeters",
-                int(math.ceil(snow_accum_cm)),
+                math.ceil(snow_accum_cm),
             ]
         elif snow_max_accum > 0:
             if snow_accum_cm == 0:
